agents/base/llamastack_agent/pyproject.toml
[tool.poetry]
name = "llamastack_agent_base"
version = "0.1.0"
description = "Agent built on LlamaStack API only (no LlamaIndex). Placeholder for full llama-stack-client integration."
authors = ["Your Name <you@example.com>"]
license = "MIT"
packages = [
    { include = "llamastack_agent_base", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.12"
python-dotenv = "^1.0.0"
llama-stack-client = ">=0.4.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
